Change GFP_DMA32 to GFP_DMA to fix an allocation failure.
Sometimes seen on bootup, the kernel panics during dma_alloc_coherent.
dma_alloc_coherent calls slab_alloc and if the freelist in slab_alloc is empty,
then new_slab would be called to create a new object. new_slab will crash the
system using BUG because it doesn't like GFP_DMA32. The flag is not checked in the
fast path, so the problem does not happen that often.