PD #73875: hdmitx: edid: optimize HDMI/DVI judgement
For most cases, we consider that the receiver is HDMI device,
such as TV or Audio Amp.
If we got an EDID that only contains one block and has valid checksum,
then we consider that is a DVI device.
If we got an EDID that all has zeroes, we consider it as a default HDMI device.

Note: if data[1] in block 1 equals to 1, we also consider it as a DVI monitor.